+++ src/ESPAsync_WiFiManager-Impl.h
@@ -14,43 +14,18 @@
   Built by Khoi Hoang https://github.com/khoih-prog/ESPAsync_WiFiManager
   Licensed under MIT license
   
-  Version: 1.12.1
+  Version: 1.12.2
 
   Version Modified By  Date      Comments
   ------- -----------  ---------- -----------
   1.0.11  K Hoang      21/08/2020 Initial coding to use (ESP)AsyncWebServer instead of (ESP8266)WebServer. Bump up to v1.0.11
                                   to sync with ESP_WiFiManager v1.0.11
-  1.1.1   K Hoang      29/08/2020 Add MultiWiFi feature to autoconnect to best WiFi at runtime to sync with
-                                  ESP_WiFiManager v1.1.1. Add setCORSHeader function to allow flexible CORS
-  1.1.2   K Hoang      17/09/2020 Fix bug in examples.
-  1.2.0   K Hoang      15/10/2020 Restore cpp code besides Impl.h code to use if linker error. Fix bug.
-  1.3.0   K Hoang      04/12/2020 Add LittleFS support to ESP32 using LITTLEFS Library
-  1.4.0   K Hoang      18/12/2020 Fix staticIP not saved. Add functions. Add complex examples.
-  1.4.1   K Hoang      21/12/2020 Fix bug and compiler warnings.
-  1.4.2   K Hoang      21/12/2020 Fix examples' bug not using saved WiFi Credentials after losing all WiFi connections.
-  1.4.3   K Hoang      23/12/2020 Fix examples' bug not saving Static IP in certain cases.
-  1.5.0   K Hoang      13/02/2021 Add support to new ESP32-S2. Optimize code.
-  1.6.0   K Hoang      25/02/2021 Fix WiFi Scanning bug.
-  1.6.1   K Hoang      26/03/2021 Modify multiWiFi-related timings to work better with latest esp32 core v1.0.6
-  1.6.2   K Hoang      08/04/2021 Fix example misleading messages.
-  1.6.3   K Hoang      13/04/2021 Allow captive portal to run more than once by closing dnsServer.
-  1.7.0   K Hoang      20/04/2021 Add support to new ESP32-C3 using SPIFFS or EEPROM
-  1.7.1   K Hoang      25/04/2021 Fix MultiWiFi bug. Fix captive-portal bug if CP AP address is not default 192.168.4.1
-  1.8.0   K Hoang      30/04/2021 Set _timezoneName. Add support to new ESP32-S2 (METRO_ESP32S2, FUNHOUSE_ESP32S2, etc.)
-  1.8.1   K Hoang      06/05/2021 Fix bug. Don't display invalid time when not synch yet.
-  1.9.0   K Hoang      08/05/2021 Add WiFi /scan page. Fix timezoneName not displayed in Info page. Clean up.
-  1.9.1   K Hoang      18/05/2021 Fix warnings with ESP8266 core v3.0.0
-  1.9.2   K Hoang      02/08/2021 Fix Mbed TLS compile error and MultiWiFi connection issue with ESP32 core v2.0.0-rc1+
-  1.9.3   K Hoang      13/08/2021 Add WiFi scanning of hidden SSIDs
-  1.9.4   K Hoang      10/10/2021 Update `platform.ini` and `library.json`
-  1.9.5   K Hoang      26/11/2021 Auto detect ESP32 core and use either built-in LittleFS or LITTLEFS library
-  1.9.6   K Hoang      26/11/2021 Fix compile error for ESP32 core v1.0.5-
-  1.9.7   K Hoang      30/11/2021 Fix bug to permit using HTTP port different from 80
-  1.9.8   K Hoang      01/12/2021 Fix bug returning IP `255.255.255.255` in core v2.0.0+ when using `hostname`
+  ...
   1.10.0  K Hoang      29/12/2021 Fix `multiple-definitions` linker error and weird bug related to src_cpp
   1.11.0  K Hoang      17/01/2022 Enable compatibility with old code to include only ESP_WiFiManager.h
   1.12.0  K Hoang      10/02/2022 Add support to new ESP32-S3
   1.12.1  K Hoang      11/02/2022 Add LittleFS support to ESP32-C3. Use core LittleFS instead of Lorol's LITTLEFS for v2.0.0+
+  1.12.2  K Hoang      13/03/2022 Optimize code by using passing by `reference` instead of by `value`
  *****************************************************************************************************************************/
 
 #pragma once
@@ -71,7 +46,8 @@ ESPAsync_WMParameter::ESPAsync_WMParameter(const char *custom)
   _customHTML = custom;
 }
 
-ESPAsync_WMParameter::ESPAsync_WMParameter(const char *id, const char *placeholder, const char *defaultValue, int length, const char *custom, int labelPlacement)
+ESPAsync_WMParameter::ESPAsync_WMParameter(const char *id, const char *placeholder, const char *defaultValue, 
+                                           const int& length, const char *custom, const int& labelPlacement)
 {
   init(id, placeholder, defaultValue, length, custom, labelPlacement);
 }
@@ -79,11 +55,13 @@ ESPAsync_WMParameter::ESPAsync_WMParameter(const char *id, const char *placehold
 // KH, using struct                      
 ESPAsync_WMParameter::ESPAsync_WMParameter(const WMParam_Data& WMParam_data)
 {
-  init(WMParam_data._id, WMParam_data._placeholder, WMParam_data._value, WMParam_data._length, "", WMParam_data._labelPlacement);
+  init(WMParam_data._id, WMParam_data._placeholder, WMParam_data._value, 
+       WMParam_data._length, "", WMParam_data._labelPlacement);
 }                  
 //////
 
-void ESPAsync_WMParameter::init(const char *id, const char *placeholder, const char *defaultValue, int length, const char *custom, int labelPlacement)
+void ESPAsync_WMParameter::init(const char *id, const char *placeholder, const char *defaultValue, 
+                                const int& length, const char *custom, const int& labelPlacement)
 {
   _WMParam_data._id = id;
   _WMParam_data._placeholder = placeholder;
@@ -121,7 +99,7 @@ void ESPAsync_WMParameter::setWMParam_Data(const WMParam_Data& WMParam_data)
   memcpy(&_WMParam_data, &WMParam_data, sizeof(_WMParam_data));
 }
 
-void ESPAsync_WMParameter::getWMParam_Data(WMParam_Data &WMParam_data)
+void ESPAsync_WMParameter::getWMParam_Data(WMParam_Data& WMParam_data)
 {
   LOGINFO(F("getWMParam_Data"));
   
@@ -1150,7 +1128,7 @@ void ESPAsync_WiFiManager::startWPS()
 
 //Convenient for debugging but wasteful of program space.
 //Remove if short of space
-const char* ESPAsync_WiFiManager::getStatus(int status)
+const char* ESPAsync_WiFiManager::getStatus(const int& status)
 {
   switch (status)
   {
@@ -1212,21 +1190,21 @@ void ESPAsync_WiFiManager::resetSettings()
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::setTimeout(unsigned long seconds)
+void ESPAsync_WiFiManager::setTimeout(const unsigned long& seconds)
 {
   setConfigPortalTimeout(seconds);
 }
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::setConfigPortalTimeout(unsigned long seconds)
+void ESPAsync_WiFiManager::setConfigPortalTimeout(const unsigned long& seconds)
 {
   _configPortalTimeout = seconds * 1000;
 }
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::setConnectTimeout(unsigned long seconds)
+void ESPAsync_WiFiManager::setConnectTimeout(const unsigned long& seconds)
 {
   _connectTimeout = seconds * 1000;
 }
@@ -1239,7 +1217,7 @@ void ESPAsync_WiFiManager::setDebugOutput(bool debug)
 //////////////////////////////////////////
 
 // KH, To enable dynamic/random channel
-int ESPAsync_WiFiManager::setConfigPortalChannel(int channel)
+int ESPAsync_WiFiManager::setConfigPortalChannel(const int& channel)
 {
   // If channel < MIN_WIFI_CHANNEL - 1 or channel > MAX_WIFI_CHANNEL => channel = 1
   // If channel == 0 => will use random channel from MIN_WIFI_CHANNEL to MAX_WIFI_CHANNEL
@@ -1302,7 +1280,7 @@ void ESPAsync_WiFiManager::setSTAStaticIPConfig(const WiFi_STA_IPConfig& WM_STA_
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::getSTAStaticIPConfig(WiFi_STA_IPConfig &WM_STA_IPconfig)
+void ESPAsync_WiFiManager::getSTAStaticIPConfig(WiFi_STA_IPConfig& WM_STA_IPconfig)
 {
   LOGINFO(F("getSTAStaticIPConfig"));
   
@@ -1327,7 +1305,7 @@ void ESPAsync_WiFiManager::setSTAStaticIPConfig(const IPAddress& ip, const IPAdd
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::setMinimumSignalQuality(int quality)
+void ESPAsync_WiFiManager::setMinimumSignalQuality(const int& quality)
 {
   _minimumQuality = quality;
 }
@@ -1341,7 +1319,7 @@ void ESPAsync_WiFiManager::setBreakAfterConfig(bool shouldBreak)
 
 //////////////////////////////////////////
 
-void ESPAsync_WiFiManager::reportStatus(String &page)
+void ESPAsync_WiFiManager::reportStatus(String& page)
 {
   page += FPSTR(WM_HTTP_SCRIPT_NTP_MSG);
 
@@ -2250,7 +2228,8 @@ void ESPAsync_WiFiManager::setSaveConfigCallback(void(*func)())
 //////////////////////////////////////////
 
 // sets a custom element to add to head, like a new style tag
-void ESPAsync_WiFiManager::setCustomHeadElement(const char* element) {
+void ESPAsync_WiFiManager::setCustomHeadElement(const char* element) 
+{
   _customHeadElement = element;
 }
 
@@ -2264,7 +2243,7 @@ void ESPAsync_WiFiManager::setRemoveDuplicateAPs(bool removeDuplicates)
 
 //////////////////////////////////////////
 
-int ESPAsync_WiFiManager::getRSSIasQuality(int RSSI)
+int ESPAsync_WiFiManager::getRSSIasQuality(const int& RSSI)
 {
   int quality = 0;